reroutes step3, step4 and step5 or profile wizard switches up appearance in profile wizard
